requirements.txt
Flask>=1.1.1
requests>=2.22.0
grequests>=0.6.0
bs4>=0.0.1
Pillow>=5.4.1
gunicorn>=20.0.4
lxml>=4.5.0
